# go/model_connection_id.go
/*
 * Сервис интерфейса P2r
 *
 * Сервис для реализации процедур интерфейса P2r. 
 *
 * API version: 1.0.
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type ConnectionId struct {
	// Идентификатор ФЧ БРТК
	FpId float64 `json:"fp_id,omitempty"`
	// Идентификатор ПММ
	RmId float64 `json:"rm_id,omitempty"`
}