# go/model_warning.go
/*
 * Сервис интерфейса P2r
 *
 * Сервис для реализации процедур интерфейса P2r. 
 *
 * API version: 1.0.
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Warning struct {
	// Идентификатор уведомления
	Id float64 `json:"id,omitempty"`
	// Время по наступления события в сек
	Time float64 `json:"time,omitempty"`
	// Причина уведомления
	Reason string `json:"reason,omitempty"`
}